# gudangku/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GudangKu",
        "description": "API Documentation for GudangKu",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/my_profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get my profile",
                "description": "This request is used to get user profile info. This request is using MySql database, and have a protected routes",
                "operationId": "c286fe042f9492a53f1a424547fa83e4",
                "responses": {
                    "200": {
                        "description": "user fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "user fetched"
                                        },
                                        "data": {
                                            "properties": {
                                                "username": {
                                                    "type": "string",
                                                    "example": "flazefy"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "flazen.edu@gmail.com"
                                                },
                                                "telegram_user_id": {
                                                    "type": "string",
                                                    "example": "1317625970"
                                                },
                                                "telegram_is_valid": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-20 22:53:47"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "telegram_data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "6f59235e-c398-8a83-2f95-3f1fbe95ca6e"
                                                },
                                                "request_type": {
                                                    "type": "string",
                                                    "example": "register"
                                                },
                                                "request_context": {
                                                    "type": "string",
                                                    "example": "OUW36L"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-20 22:53:47"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "protected route need to include sign in token as authorization bearer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "you need to include the authorization token from login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user failed to fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "user not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}